4.20-4.21

+++ viikko4/controllers/blogs.js
@@ -3,14 +3,6 @@ const jwt = require('jsonwebtoken');
 const Blog = require('../models/blog');
 const User = require('../models/user');
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization');
-  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-    return authorization.substring(7);
-  }
-  return null;
-};
-
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1 });
   response.json(blogs);
@@ -20,7 +12,7 @@ blogsRouter.post('/', async (request, response, next) => {
   const body = request.body;
   console.log(body);
 
-  const token = getTokenFrom(request);
+  const token = request.token;
 
   try {
     const decodedToken = jwt.verify(token, process.env.SECRET);
@@ -52,9 +44,22 @@ blogsRouter.post('/', async (request, response, next) => {
 });
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const token = request.token;
+
   try {
-    await Blog.findByIdAndRemove(request.params.id);
-    response.status(204).end();
+    const decodedToken = jwt.verify(token, process.env.SECRET);
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' });
+    }
+
+    const blog = await Blog.findById(request.params.id);
+
+    if (blog.user.toString() === decodedToken.id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id);
+      response.status(204).end();
+    }
+
+    response.status(400).send({ error: 'you can only remove your own blogs' });
   } catch (exception) {
     next(exception);
   }
